Use new YAML parser for config
