+++ src/main/java/org/lavajuno/mirrorlog/yaml/YamlList.java
@@ -1,40 +1,44 @@
-package org.lavajuno.mirrorlog.yaml2;
+package org.lavajuno.mirrorlog.yaml;
 
-import java.util.*;
+import java.util.InvalidPropertiesFormatException;
+import java.util.Vector;
+import java.util.List;
 
 /**
- * YamlMap represents an element containing an unordered map of elements.
+ * Represents an element containing an ordered list of elements.
  */
-public class YamlMap extends YamlElement {
+public class YamlList extends YamlElement {
 
     protected final String KEY;
 
-    protected final TreeMap<String, YamlElement> ELEMENTS;
+    protected final Vector<YamlElement> ELEMENTS;
 
-    protected YamlMap(String key, List<String> lines, int begin, Integer end, int indent) throws InvalidPropertiesFormatException {
+    protected YamlList(String key, List<String> lines, int begin, Integer end, int indent) throws InvalidPropertiesFormatException {
         this.KEY = key;
-        this.ELEMENTS = new TreeMap<>();
+        this.ELEMENTS = new Vector<>();
+
         for(int i = begin; i < lines.size(); i++) {
             String line = lines.get(i);
             int line_indent = parseIndent(line);
             LINE_MATCHES line_match = matchLine(line);
 
             if(line_indent == indent) { /* If indent matches */
-                String curr_key;
                 switch(line_match) {
-                    case ELEMENT:
-                        curr_key = parseKey(line);
-                        ELEMENTS.put(curr_key, parseElement(curr_key, lines, i, end, indent + 2));
+                    case LIST_ELEMENT:
+                        String curr_key = parseKey(line.split("-", 2)[1]);
+                        ELEMENTS.add(parseElement(curr_key, lines, i, end, indent + 2));
                         break;
-                    case PAIR:
-                        curr_key = parseKey(line);
-                        ELEMENTS.put(curr_key, new YamlPair(line));
+                    case LIST_PAIR:
+                        ELEMENTS.add(new YamlPair(line.split("-", 2)[1].stripLeading()));
+                        break;
+                    case LIST_VALUE:
+                        ELEMENTS.add(new YamlValue(line.split("-", 2)[1].stripLeading()));
                         break;
                     case IGNORE:
                         break;
                     default:
                         /* If no match was found, we don't know what to do with the line */
-                        printParseError(line, i, "Unexpected element. (Not part of an unordered map)");
+                        printParseError(line, i, "Unexpected element. (Not part of an ordered list)");
                         throw new InvalidPropertiesFormatException("Parse error on line " + i + " of input.");
                 }
             } else if(line_indent < indent && line_match != LINE_MATCHES.IGNORE) {
@@ -44,25 +48,21 @@ public class YamlMap extends YamlElement {
         }
     }
 
-    public YamlMap(List<String> lines) throws InvalidPropertiesFormatException {
-        this("root", lines, 0, 0, 0);
-    }
-
     public String getKey() { return KEY; }
 
-    public Map<String, YamlElement> getElements() { return ELEMENTS; }
+    public List<YamlElement> getElements() { return ELEMENTS; }
 
-    public Set<String> getKeys() { return ELEMENTS.keySet(); }
+    public YamlElement getElement(int index) { return ELEMENTS.get(index); }
 
-    public YamlElement getElement(String key) { return ELEMENTS.get(key); }
+    public int getSize() { return ELEMENTS.size(); }
 
     protected String toString(int indent, boolean list) {
         StringBuilder sb = new StringBuilder();
         sb.append(" ".repeat(indent));
         if(list) { sb.append("- "); }
         sb.append(KEY).append(":\n");
-        for(String i : ELEMENTS.keySet()) {
-            sb.append(ELEMENTS.get(i).toString(indent + 2, false));
+        for(YamlElement i : ELEMENTS) {
+            sb.append(i.toString(indent + 2, true));
         }
         return sb.toString();
     }
