+++ src/main/java/org/lavajuno/mirrorlog/yaml/YamlMap.java
@@ -1,44 +1,40 @@
-package org.lavajuno.mirrorlog.yaml2;
+package org.lavajuno.mirrorlog.yaml;
 
-import java.util.InvalidPropertiesFormatException;
-import java.util.Vector;
-import java.util.List;
+import java.util.*;
 
 /**
- * Represents an element containing an ordered list of elements.
+ * YamlMap represents an element containing an unordered map of elements.
  */
-public class YamlList extends YamlElement {
+public class YamlMap extends YamlElement {
 
     protected final String KEY;
 
-    protected final Vector<YamlElement> ELEMENTS;
+    protected final TreeMap<String, YamlElement> ELEMENTS;
 
-    protected YamlList(String key, List<String> lines, int begin, Integer end, int indent) throws InvalidPropertiesFormatException {
+    protected YamlMap(String key, List<String> lines, int begin, Integer end, int indent) throws InvalidPropertiesFormatException {
         this.KEY = key;
-        this.ELEMENTS = new Vector<>();
-
+        this.ELEMENTS = new TreeMap<>();
         for(int i = begin; i < lines.size(); i++) {
             String line = lines.get(i);
             int line_indent = parseIndent(line);
             LINE_MATCHES line_match = matchLine(line);
 
             if(line_indent == indent) { /* If indent matches */
+                String curr_key;
                 switch(line_match) {
-                    case LIST_ELEMENT:
-                        String curr_key = parseKey(line.split("-", 2)[1]);
-                        ELEMENTS.add(parseElement(curr_key, lines, i, end, indent + 2));
+                    case ELEMENT:
+                        curr_key = parseKey(line);
+                        ELEMENTS.put(curr_key, parseElement(curr_key, lines, i, end, indent + 2));
                         break;
-                    case LIST_PAIR:
-                        ELEMENTS.add(new YamlPair(line.split("-", 2)[1].stripLeading()));
-                        break;
-                    case LIST_VALUE:
-                        ELEMENTS.add(new YamlValue(line.split("-", 2)[1].stripLeading()));
+                    case PAIR:
+                        curr_key = parseKey(line);
+                        ELEMENTS.put(curr_key, new YamlPair(line));
                         break;
                     case IGNORE:
                         break;
                     default:
                         /* If no match was found, we don't know what to do with the line */
-                        printParseError(line, i, "Unexpected element. (Not part of an ordered list)");
+                        printParseError(line, i, "Unexpected element. (Not part of an unordered map)");
                         throw new InvalidPropertiesFormatException("Parse error on line " + i + " of input.");
                 }
             } else if(line_indent < indent && line_match != LINE_MATCHES.IGNORE) {
@@ -48,21 +44,25 @@ public class YamlList extends YamlElement {
         }
     }
 
+    public YamlMap(List<String> lines) throws InvalidPropertiesFormatException {
+        this("root", lines, 0, 0, 0);
+    }
+
     public String getKey() { return KEY; }
 
-    public List<YamlElement> getElements() { return ELEMENTS; }
+    public Map<String, YamlElement> getElements() { return ELEMENTS; }
 
-    public YamlElement getElement(int index) { return ELEMENTS.get(index); }
+    public Set<String> getKeys() { return ELEMENTS.keySet(); }
 
-    public int getSize() { return ELEMENTS.size(); }
+    public YamlElement getElement(String key) { return ELEMENTS.get(key); }
 
     protected String toString(int indent, boolean list) {
         StringBuilder sb = new StringBuilder();
         sb.append(" ".repeat(indent));
         if(list) { sb.append("- "); }
         sb.append(KEY).append(":\n");
-        for(YamlElement i : ELEMENTS) {
-            sb.append(i.toString(indent + 2, true));
+        for(String i : ELEMENTS.keySet()) {
+            sb.append(ELEMENTS.get(i).toString(indent + 2, false));
         }
         return sb.toString();
     }
